Dirty hacks to get selection working for exsting values.
For some reason this didn't seem to work out of the box for me. I'm
a rails n00b so don't know much about its internals, but from what
I can tell the library was only really using object_name for the
FormBuilder which didn't have access to the current value.

I've basically patched this the quickest way I could to make the object
itself available.